feat(viewer): every click does something — hide a node that has nothing to expand or collapse (#76)
Users reported that seed nodes with no hidden neighbors and no
dependents did nothing on click, which felt broken. Introduce
Graph.Shaping.shrink and a HideNode action: clicking such a node now
removes it from the visible set (overriding InitialSeed protection,
since the click is an explicit user request). Removes any other node
whose only reason was being anchored by the hidden node.

Click cycle on any visible node:
  has hidden neighbors   → expand
  has anchor dependents  → collapse them
  neither                → hide this node

Also prunes the stale diagnose.spec.ts and the obsolete
expand-collapse.spec.ts (superseded by click-to-expand.spec.ts and
centric-layout.spec.ts).
