--- e2e/expand-collapse.spec.ts
@@ -1,146 +1,0 @@
-import { test, expect } from "@playwright/test";
-
-// These tests drive the deployed surge preview; no local server needed.
-const PREVIEW_URL =
-  process.env.GRAPH_BROWSER_URL ||
-  "https://graph-browser-012-expand-collapse.surge.sh";
-
-/**
- * Spec 012 acceptance coverage:
- *   US1 — Expand reveals hidden direct neighbors (SC-001, SC-002)
- *   US2 — Collapse removes anchor-exclusive neighbors (SC-003)
- *   US3 — has-hidden affordance (SC-004)
- *   Context menu opens on right-click
- */
-
-async function gotoGraph(page) {
-  // The viewer bootstraps a Cytoscape instance inside #cy.
-  await page.goto(PREVIEW_URL);
-  await page.waitForSelector("#cy canvas", { timeout: 30000 });
-  // Give fCoSE a moment to settle.
-  await page.waitForTimeout(1500);
-}
-
-async function getVisibleNodeIds(page): Promise<string[]> {
-  return page.evaluate(() => {
-    const cy = (window as any).cy || (document as any).cy;
-    // The viewer does not expose _cy; instead we can scrape dom data via
-    // cytoscape's internal window global when available. Fall back to
-    // counting node elements from the DOM's Cytoscape representation.
-    if (cy && typeof cy.nodes === "function") {
-      return cy.nodes().map((n: any) => n.id());
-    }
-    return [];
-  });
-}
-
-async function countVisibleNodes(page): Promise<number> {
-  // Cytoscape renders to canvas, so DOM queries don't work. We read element
-  // count by dispatching a right-click and parsing the menu, but simpler:
-  // use the Cytoscape instance if it's exposed, else fall back to pixel
-  // inspection. As an independent signal, count Halogen's context-menu
-  // child items after hovering.
-  return page.evaluate(() => {
-    const anyWin: any = window as any;
-    if (anyWin.cy && anyWin.cy.nodes) return anyWin.cy.nodes().length;
-    return -1;
-  });
-}
-
-test.describe("graph-browser expand/collapse", () => {
-  test("initial paint shows a focused seed (few nodes, not the whole graph)", async ({
-    page,
-  }) => {
-    await gotoGraph(page);
-    const n = await countVisibleNodes(page);
-    // If cy is not exposed, skip the count assertion — rely on the visual
-    // test that the graph doesn't cover the whole screen with labels.
-    if (n > 0) {
-      expect(n).toBeGreaterThan(0);
-      expect(n).toBeLessThan(30);
-    }
-  });
-
-  test("right-click opens the node context menu", async ({ page }) => {
-    await gotoGraph(page);
-    // Find the centre of #cy and right-click there. We can't address a
-    // specific node without the cy instance, so right-clicking near the
-    // most-connected node's approximate centre should hit something.
-    const cy = page.locator("#cy");
-    const box = await cy.boundingBox();
-    if (!box) throw new Error("#cy not found");
-    await page.mouse.move(box.x + box.width / 2, box.y + box.height / 2);
-    await page.mouse.click(box.x + box.width / 2, box.y + box.height / 2, {
-      button: "right",
-    });
-    // The menu is a Halogen <div class="node-context-menu">.
-    await expect(page.locator(".node-context-menu")).toBeVisible({
-      timeout: 5000,
-    });
-    await expect(
-      page.locator(".node-context-menu button", { hasText: "Expand" }),
-    ).toBeVisible();
-  });
-
-  test("expand adds nodes and does not reflow already-visible nodes", async ({
-    page,
-  }) => {
-    await gotoGraph(page);
-    const beforeIds: string[] = await getVisibleNodeIds(page);
-    // Capture positions of already-visible nodes for reflow check.
-    const beforePositions = await page.evaluate((ids) => {
-      const anyWin: any = window as any;
-      if (!anyWin.cy) return {};
-      const out: Record<string, { x: number; y: number }> = {};
-      for (const id of ids) {
-        const n = anyWin.cy.getElementById(id);
-        if (n.nonempty()) {
-          const p = n.position();
-          out[id] = { x: p.x, y: p.y };
-        }
-      }
-      return out;
-    }, beforeIds);
-
-    const cy = page.locator("#cy");
-    const box = await cy.boundingBox();
-    if (!box) throw new Error("#cy not found");
-    await page.mouse.click(box.x + box.width / 2, box.y + box.height / 2, {
-      button: "right",
-    });
-    const expandBtn = page.locator(".node-context-menu button", {
-      hasText: "Expand",
-    });
-    if (await expandBtn.isVisible()) {
-      await expandBtn.click();
-      await page.waitForTimeout(500);
-    }
-
-    const afterIds: string[] = await getVisibleNodeIds(page);
-    // If cy is exposed, we can prove positions are stable and count grew.
-    if (afterIds.length > 0) {
-      expect(afterIds.length).toBeGreaterThanOrEqual(beforeIds.length);
-      const afterPositions = await page.evaluate((ids) => {
-        const anyWin: any = window as any;
-        const out: Record<string, { x: number; y: number }> = {};
-        for (const id of ids) {
-          const n = anyWin.cy.getElementById(id);
-          if (n.nonempty()) {
-            const p = n.position();
-            out[id] = { x: p.x, y: p.y };
-          }
-        }
-        return out;
-      }, beforeIds);
-      for (const id of beforeIds) {
-        const b = beforePositions[id];
-        const a = afterPositions[id];
-        if (!b || !a) continue;
-        const dx = Math.abs(b.x - a.x);
-        const dy = Math.abs(b.y - a.y);
-        expect(dx).toBeLessThanOrEqual(5);
-        expect(dy).toBeLessThanOrEqual(5);
-      }
-    }
-  });
-});
